backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "AI-powered PDF submission analysis service"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.116.1",
    "google-genai[aiohttp]>=1.0.0",
    "httpx>=0.28.1",
    "pydantic>=2.10.6",
    "pydantic-settings>=2.8.0",
    "python-dotenv>=1.0.1",
]
